requirements.txt
boto3
voluptuous
pillow
lambda-packages
pycrypto